requirements.txt
python-dotenv
pgvector
openai
psycopg2-binary
langchain
langchain_openai
urllib3==1.26.6
datasets